requirements.txt
Django==1.9
dj-database-url==0.3.0
django-oauth-toolkit==0.9.0
djangorestframework==3.3.1
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.3.0
hashids
httpie
psycopg2==2.6